HDLcd: Enable 4K resolution for gem5
This patch adds a 4K resolution option to the HDLcd controller
which the real hardware is not capable of. This mode will be useful
for future-looking studies. The default resolution is unchanged;
you will need to update the dts file to enable 4K (see the new
comment in the dts file HDLcd controller section).  You will also
want to change the HDLcd oscillator frequency in RealView.py to get
the frame rate you want as writing of the Versatile Express
oscillator registers currently doesn't have any effect on the gem5
model.  Tested with Android Jelly Bean.

Signed-off-by: Christopher Daniel Emmons <chris.emmons@arm.com>