gem5: Add support for gem5's extended GIC mode
gem5's GIC model contains a gem5-specific extension that enables
support for up to 255 CPUs. When enabled, the GIC interprets registers
where a CPU bitmask would normally be present as a CPU ID.

This extensions need to be enabled by setting the OF compatible string
to "gem5,gic".

Kudos to Matt Evans for the original version of this patch.

Change-Id: I1a8776f1adcbfcd97813573050a8595176de6a7e
Signed-off-by: Andreas Sandberg <andreas.sandberg@arm.com>
Reviewed-by: Stephan Diestelhorst <stephan.diestelhorst@arm.com>
Reviewed-by: Curtis Dunham <curtis.dunham@arm.com>
[Ported to 4.4]
Signed-off-by: Andreas Sandberg <andreas.sandberg@arm.com>