gem5, ARM64: Enable the CPUFreq framework and corresponding drivers
In this changeset we enable the CPUFreq framework, as well as the
corresponding gem5 drivers, in the gem5_defconfig for arm64. We enable
the following goverors:

* Performance (Default)
* Powersave
* Userspace
* Ondemand
* Conservative

The default governor is Performance as this will not adjust the
frequency, and will leave it at the maximum frequency. Therefore,
unless the governor is changed, future simulations should run as
before.

Change-Id: Iffe1734b92cc0421c2844b0e784b0ba34143d28d
Signed-off-by: Sascha Bischoff <sascha.bischoff@arm.com>
Reviewed-by: Geoffrey Blake <geoffrey.blake@arm.com>
Signed-off-by: Andreas Sandberg <andreas.sandberg@arm.com>